requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-debug-toolbar==2.2
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.11.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
MarkupSafe==1.1.1
mysqlclient==1.4.6
Pillow==7.0.0
python-memcached==1.59
pytz==2017.2
redis==3.4.1
requests==2.22.0
shortuuid==0.5.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
Werkzeug==1.0.0
